Added per-tape nozzle selection, added error status for feeders, various bug fixes
Each tape is now associated with a nozzle and the software will
automatically change nozzle and recalibrate if nozzle changer is
enabled. A new "ERROR" status is added for feeders that have failed
pickup 3 times in a row, this disables the feeder and will have to be
re-enabled by the user. Quit menu now quits the program. TinyG is now
reset at startup in order to avoid arbitrary movements at startup. Added
checks for null pointers. Z axis not able to jog below table height
anymore. If Auto zero of up camera fails a message is shown.